Extract solve's success-path cleanup so the file stays under the line threshold
Adding docker-image reclamation to cleanupTempDirectory pushed
src/solve.repository.lib.mjs from 1346 to 1387 lines, back over the
1350-line early-warning threshold that scripts/check-file-line-limits.sh
enforces (#1593) and that tests/extracted-modules-2198.test.mjs pins.

The three success-path helpers - workspace, agent snapshot stores (#2186)
and superseded docker images (#2187) - run after the work is done and
share nothing with the repository-setup flow around them, so they move to
src/solve.cleanup.lib.mjs following the precedent of #2175 and #2198. All
three names stay re-exported, so no importer changes.

tests/test-issue-2186-agent-snapshot-reclaim.mjs reads that source to
check the wiring and now reads the new module. Its ordering assertion
searched for 'shouldCleanup', which has never been the name of the
keep/delete branch ('shouldKeepDirectory'), so indexOf returned -1 and
the check could not fail; it now matches the real identifier.
